Add an fbdev screen as a fallback on sparc
Because of a bug in our PCI code that got exposed by recent kernels,
we will fail to start with PCI drivers on sparc.  Add a new fallback
screen using the fbdev driver that should hopefully still work.

For this to work we need xf86MapDomainMemory to not throw a fatal error
on failure, otherwise we won't even try the fallback.